Scroll to top, remove old editors, remove console.logs
Three things in this one commit. (1) After snippets update, we use $anchorScroll
to scroll to the top of the page. (2) Before new editors are loaded onto the
page, the old editors are removed. It seemed like things we getting laggy and
I believe this was the cause. (3) Remove console.log statements.